requirements.txt
Flask==2.3.2
ccxt==4.2.96
python-dotenv==1.0.1
